Pulsar/requirements.txt
pulsar-client==2.1.1
blockade==0.4.0